refactor: flatten job executor
