+++ internal/app/api/v1/tests.go
@@ -12,11 +12,11 @@ import (
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
 	"github.com/kubeshop/testkube/pkg/crd"
 	"github.com/kubeshop/testkube/pkg/cronjob"
+	"github.com/kubeshop/testkube/pkg/executor/client"
 	testsmapper "github.com/kubeshop/testkube/pkg/mapper/tests"
 	"github.com/kubeshop/testkube/pkg/secret"
 	"go.mongodb.org/mongo-driver/mongo"
 
-	"github.com/kubeshop/testkube/pkg/jobs"
 	"k8s.io/apimachinery/pkg/api/errors"
 )
 
@@ -491,10 +491,10 @@ func (s TestkubeAPI) DeleteTestsHandler() fiber.Handler {
 
 func GetSecretsStringData(content *testkube.TestContent) map[string]string {
 	// create secrets for test
-	stringData := map[string]string{jobs.GitUsernameSecretName: "", jobs.GitTokenSecretName: ""}
+	stringData := map[string]string{client.GitUsernameSecretName: "", client.GitTokenSecretName: ""}
 	if content != nil && content.Repository != nil {
-		stringData[jobs.GitUsernameSecretName] = content.Repository.Username
-		stringData[jobs.GitTokenSecretName] = content.Repository.Token
+		stringData[client.GitUsernameSecretName] = content.Repository.Username
+		stringData[client.GitTokenSecretName] = content.Repository.Token
 	}
 
 	return stringData
